2d_recognition/requirements.txt
absl-py==0.7.1
astor==0.7.1
cycler==0.10.0
gast==0.2.2
grpcio==1.20.0
h5py==2.9.0
idx2numpy==1.2.2
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.1
matplotlib==3.0.3
mock==2.0.0
numpy==1.16.2
opencv-python==4.1.0.25
pbr==5.1.3
protobuf==3.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow-gpu==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==0.15.2
